Factor out common code across server and nl_server
