+++ nl_server/requirements.txt
@@ -6,7 +6,7 @@ google-cloud-pubsub==2.14.0
 google-cloud-storage==2.4.0
 gunicorn==20.1.0
 markupsafe==2.0.1
-pandas==1.3.5
+pandas>=1.3.5
 sentence_transformers==2.2.2
 spacy==3.5.0
 pydantic==1.9.2
