+++ server/services/nl.py
@@ -18,7 +18,6 @@ import logging
 import re
 from typing import Dict, List, Union
 
-import server.lib.nl.constants as constants
 from server.lib.nl.detection import BinaryClassificationResultType
 from server.lib.nl.detection import ClassificationType
 from server.lib.nl.detection import ComparisonClassificationAttributes
@@ -36,14 +35,15 @@ from server.lib.nl.detection import SizeTypeClassificationAttributes
 from server.lib.nl.detection import TimeDeltaClassificationAttributes
 from server.lib.nl.detection import TimeDeltaType
 from server.lib.nl.place_detection import NLPlaceDetector
-import server.lib.nl.utils as utils
 from server.services import datacommons as dc
+import shared.lib.constants as constants
+import shared.lib.utils as shared_utils
 
 # TODO: decouple words removal from detected attributes. Today, the removal
 # blanket removes anything that matches, including the various attribute/
 # classification triggers and contained_in place types (and their plurals).
 # This may not always be the best thing to do.
-ALL_STOP_WORDS = utils.combine_stop_words()
+ALL_STOP_WORDS = shared_utils.combine_stop_words()
 
 
 def pick_best(probs):
@@ -67,17 +67,6 @@ def pick_option(class_model, q,
     return categories[class_model.predict([q])[0]]
   else:
     return None
-
-
-def _prefix_length(s1, s2):
-  if not s1 or not s2:
-    return 0
-  short_str = min([s1, s2], key=len)
-  for i, char in enumerate(short_str):
-    for other in [s1, s2]:
-      if other[i] != char:
-        return i + 1
-  return len(short_str)
 
 
 class Model:
@@ -408,7 +397,7 @@ class Model:
     logging.info(f"SV Detection: Query provided to SV Detection: {query}")
     debug_logs["sv_detection_query_input"] = query
     debug_logs["sv_detection_query_stop_words_removal"] = \
-        utils.remove_stop_words(query, ALL_STOP_WORDS)
+        shared_utils.remove_stop_words(query, ALL_STOP_WORDS)
 
     # Make API call to the NL models/embeddings server.
     return dc.nl_search_sv(query)
